fixed bug with input error message, added scss and color variables, moved createStor to stor folder, moved types to store folder
