+++ src/components/Modals/Modals.scss
@@ -1,3 +1,5 @@
+@import "../../styles/colors.scss";
+
 /* ---------- MEDIA SCREEN -- MOBILE ---------- */
 @media (max-width: 600px) {
   .modal__body {
@@ -32,7 +34,7 @@
   right: 0;
   bottom: 0;
   left: 0;
-  background: lightgray;
+  background: $background_color_lightgrey;
   display: flex;
   justify-content: center;
   overflow-y: scroll;
@@ -40,11 +42,11 @@
 
 .modal__body {
   padding: 7vmin;
-  background: white;
+  background: $background_color_white;
 }
 
 .modal__title {
-  color: #f57f17;
+  color: $text_color_orange;
   text-align: center;
   font-size: 18px;
 }
@@ -61,7 +63,7 @@
   font-weight: bold;
   margin: 5px;
   width: 100%;
-  color: grey;
+  color: $text_color_grey;
 }
 
 .modal__input {
@@ -69,11 +71,11 @@
   outline: none;
   padding: 5px;
   border-radius: 3px;
-  border: 1px solid grey;
+  border: 1px solid $border_color_grey;
 }
 
 .modal__datetime {
-  border: 1px solid lightgrey;
+  border: 1px solid $border_color_light;
   border-radius: 5px;
   padding: 5px 10px;
   width: 100%;
@@ -83,15 +85,15 @@
   margin-left: 5px;
   padding: 0px 16px;
   border-radius: 5px;
-  border: 2px solid #f57f17;
-  background-color: #f57f17;
-  color: white;
+  border: 2px solid $border_color_orange;
+  background-color: $background_color_orange;
+  color: $text_color_white;
   font-size: 24px;
 }
 
 .modal__btn-add:hover {
-  background-color: white;
-  color: #f57f17;
+  background-color: $background_color_white;
+  color: $text_color_orange;
 }
 
 .modal__controls-wrapper {
@@ -104,8 +106,8 @@
 .modal__controls-btn {
   padding: 0px 16px;
   border-radius: 5px;
-  border: 2px solid #f57f17;
-  background-color: #f57f17;
-  color: white;
+  border: 2px solid $border_color_orange;
+  background-color: $background_color_orange;
+  color: $text_color_white;
   font-size: 18px;
 }
