--- src/components/Todos/Todos.css
@@ -1,24 +1,0 @@
-.todo-item__wrapper {
-  display: flex;
-  justify-content: space-between;
-  list-style: none;
-  border: 1px solid lightgray;
-  border-radius: 5px;
-  padding: 5px 10px;
-  margin: 10px 0;
-}
-
-.todo-item {
-  display: flex;
-  align-items: center;
-}
-
-.todo-item__dates {
-  display: flex;
-  flex-direction: column;
-  margin-right: 20px;
-}
-
-.todo-items__empty-page {
-  text-align: center;
-}
