+++ src/containers/Footer/Footer.js
@@ -1,4 +1,4 @@
-import "./Footer.css";
+import "./Footer.scss";
 
 export default function Footer() {
   return (
