+++ src/index.js
@@ -1,14 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
-import { composeWithDevTools } from "redux-devtools-extension";
-import { createStore } from "redux";
 
 import "./styles/cssReset.css";
 import App from "./App";
-import rootReducer from "./store/reducers";
-
-const store = createStore(rootReducer, composeWithDevTools());
+import store from "./store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
