--- src/store/reducers.js
@@ -1,8 +1,0 @@
-import { combineReducers } from "redux";
-import todosReducer from "./todos/reducers";
-
-const rootReducer = combineReducers({
-  todos: todosReducer,
-});
-
-export default rootReducer;
