+++ src/store/todos/actions.js
@@ -1,3 +1,3 @@
-import { ADD_TODO } from "../../types/todos";
+import { ADD_TODO } from "./../types/todos";
 
 export const createTodo = (todo) => ({ type: ADD_TODO, payload: todo });
