+++ src/styles/colors.scss
@@ -1,0 +1,16 @@
+/* ---------- TEXT ---------- */
+$text_color_orange: #f57f17;
+$text_color_black: black;
+$text_color_grey: grey;
+$text_color_white: white;
+
+/* ---------- BORDER ---------- */
+$border_color_grey: grey;
+$border_color_light: lightgray;
+$border_color_orange: #f57f17;
+
+/* ---------- BACKGROUND ---------- */
+$background_color_error: blanchedalmond;
+$background_color_lightgrey: lightgray;
+$background_color_white: white;
+$background_color_orange: #f57f17;
